Workaround: "Pause" & "Sync" menu item doesn't update on Ubuntu
In ubuntu, update UI event is not invoked. Because of this GoSync
is not able to dynamically change the menu item text.

The workaround is to break the status bar in two and show the
current status of sync in the right hand side bar.

Menu item reads an static text: Pause/Resume Sync

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>