Fast start on cached data
Everytime GoSync comes up, it calculates the drive usage.
Now save the calculated data in gosyncrc and read/display
it on next boot.

The cached data will be updated with a fresh fetch when
a file is downloaded.

Its expensive right now because it queries the complete
drive and recalculates everything. Next versions will
fix this.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>