Seperated drive tree retrival from the drive usage calculation.
Earlier when settings option was opened, it used to take time
to retrive the drive tree because the drive tree retrival also
used to calculate the drive usage which can take lot of time.
Now device tree retrival is done in seperate thread and its
lot faster.

Also make other cleanups like removed the locking when calculating
drive usage. With this the drive usage might be slightly off
the actual usage but next time when GoSync comes up, it should
calculate is just fine.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>