Issue #23
If an Upload/Download fails, the exception is not handled.
This causes the subroutine to exit without releasing the
sync_lock, causing the program to freeze indefinitly.

This fixes another problem of file of zero size files
upload which causes the freeze.

Other problem that this fixes is that after sometime if
the sync is started, during validate settings the connection
is reset by peer. Now we handle this situation more gracefully
and retry the connection.

Signed-off-by: Himanshu Chauhan <hschauhan@nulltrace.org>